requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.0
flask-restx>=0.5.1
marshmallow>=3.0
Pillow>=9.0
python-dotenv>=1.0
psycopg2-binary>=2.9